refactor: split sync, verify, config, and CLI command handling
Behavior-preserving continuation of the module split (after the LanceDB
split):

- sync/ -> mod.rs (orchestration: lock, txn, manifest diff, cursor),
  indexer.rs (per-file read/chunk/embed/sweep), result.rs (accounting)
- verify/ -> mod.rs (verify orchestration + stale repair), status.rs,
  health.rs (check), sampling.rs (expected-doc-id sampling)
- config/ -> model.rs (data model + defaults), mod.rs (load/save + tests)
- main.rs trimmed to a thin binary; command handlers moved to
  src/commands.rs in the library

Extracted the duplicated doc-id derivation (sync_file vs verify
sampling) into id::doc_ids_for_chunks with focused tests for duplicate
chunk handling and determinism. Manual QA: full CLI flow (init, sync
--full, status, check, query, verify --fix, dry-run, incremental sync,
--help) produces identical output to pre-refactor.

Closes #22
