+++ src/lib.rs
@@ -1,5 +1,6 @@
 pub mod chunker;
 pub mod cli;
+pub mod commands;
 pub mod config;
 pub mod embedder;
 pub mod error;
