+++ src/sync/mod.rs
@@ -1,27 +1,25 @@
+//! Incremental sync orchestration: locking, transaction lifecycle, manifest
+//! diffing, and cursor advancement. Per-file work lives in [`indexer`];
+//! result accounting helpers live in [`result`].
+
+mod indexer;
+mod result;
+
 use crate::chunker::Chunker;
 use crate::config::Config;
 use crate::embedder::Embedder;
 use crate::error::{MinSyncError, Result};
-use crate::id::{compute_doc_id, content_hash};
 use crate::manifest::{FileChange, Manifest};
-use crate::normalize::normalize_text;
 use crate::state::{Cursor, FileLock, Transaction};
 use crate::types::SyncResult;
-use crate::vectorstore::{Document, DocumentUpdate, Filter, VectorStore};
-use std::collections::HashMap;
+use crate::vectorstore::{Filter, VectorStore};
+use indexer::{index_file, SyncFileContext};
+use result::{change_path, empty_sync_result};
 use std::path::PathBuf;
 
 pub struct MinSync {
     root: PathBuf,
     minsync_dir: PathBuf,
-}
-
-struct SyncFileContext<'a> {
-    config: &'a Config,
-    chunker: &'a dyn Chunker,
-    embedder: &'a dyn Embedder,
-    store: &'a mut dyn VectorStore,
-    sync_token: &'a str,
 }
 
 impl MinSync {
@@ -106,19 +104,7 @@ impl MinSync {
         .save(&txn_path)?;
 
         let start = std::time::Instant::now();
-        let mut result = SyncResult {
-            files_processed: 0,
-            chunks_added: 0,
-            chunks_updated: 0,
-            chunks_deleted: 0,
-            dry_run: false,
-            already_up_to_date: false,
-            files_processed_paths: Vec::new(),
-            elapsed_seconds: 0.0,
-            embedding_api_calls: 0,
-            embedded_texts: 0,
-            estimated_tokens: 0,
-        };
+        let mut result = empty_sync_result(false, false);
 
         for change in &changes {
             match change {
@@ -130,7 +116,7 @@ impl MinSync {
                         store,
                         sync_token: &sync_token,
                     };
-                    self.sync_file(path, context, &mut result).await?;
+                    index_file(&self.root, path, context, &mut result).await?;
                 }
                 FileChange::Deleted(path) => {
                     let deleted = store.delete_by_filter(&Filter::And(vec![
@@ -159,122 +145,6 @@ impl MinSync {
 
         result.elapsed_seconds = start.elapsed().as_secs_f64();
         Ok(result)
-    }
-
-    async fn sync_file(
-        &self,
-        path: &str,
-        context: SyncFileContext<'_>,
-        result: &mut SyncResult,
-    ) -> Result<()> {
-        let raw_text = match std::fs::read_to_string(self.root.join(path)) {
-            Ok(text) => text,
-            Err(error) if error.kind() == std::io::ErrorKind::InvalidData => String::new(),
-            Err(error) => return Err(MinSyncError::Io(error)),
-        };
-        let text = normalize_text(&raw_text, &context.config.normalize);
-        let chunks = context.chunker.chunk(&text, path)?;
-        let schema_id = context.chunker.schema_id();
-        let mut docs_to_embed = Vec::new();
-        let mut dup_counter: HashMap<String, usize> = HashMap::new();
-
-        for chunk in chunks {
-            let chunk_content_hash = content_hash(&chunk.text);
-            let dup_key = format!("{}\0{}", chunk_content_hash, chunk.heading_path);
-            let dup_index = dup_counter.entry(dup_key).or_insert(0);
-            let doc_id = compute_doc_id(
-                &context.config.source_id,
-                path,
-                schema_id,
-                &chunk.chunk_type,
-                &chunk_content_hash,
-                *dup_index,
-            );
-            *dup_index += 1;
-
-            if context
-                .store
-                .fetch(std::slice::from_ref(&doc_id))?
-                .is_empty()
-            {
-                docs_to_embed.push(Document {
-                    id: doc_id,
-                    embedding: Vec::new(),
-                    text: chunk.text,
-                    source_id: context.config.source_id.clone(),
-                    path: path.to_string(),
-                    chunk_schema_id: schema_id.to_string(),
-                    chunk_type: chunk.chunk_type,
-                    heading_path: chunk.heading_path,
-                    content_hash: chunk_content_hash,
-                    seen_token: context.sync_token.to_string(),
-                });
-            } else {
-                context.store.update(&[DocumentUpdate {
-                    id: doc_id,
-                    seen_token: context.sync_token.to_string(),
-                    path: path.to_string(),
-                    heading_path: chunk.heading_path,
-                }])?;
-                result.chunks_updated += 1;
-            }
-        }
-
-        if !docs_to_embed.is_empty() {
-            let texts: Vec<String> = docs_to_embed.iter().map(|doc| doc.text.clone()).collect();
-            let embeddings = context.embedder.embed(&texts).await?;
-            if embeddings.len() != docs_to_embed.len() {
-                return Err(MinSyncError::Embedding(format!(
-                    "expected {} embeddings, got {}",
-                    docs_to_embed.len(),
-                    embeddings.len()
-                )));
-            }
-
-            result.embedding_api_calls += 1;
-            result.embedded_texts += texts.len();
-            result.estimated_tokens += texts.iter().map(|text| text.len() / 4).sum::<usize>();
-
-            for (doc, embedding) in docs_to_embed.iter_mut().zip(embeddings) {
-                doc.embedding = embedding;
-            }
-            result.chunks_added += docs_to_embed.len();
-            context.store.upsert(&docs_to_embed)?;
-        }
-
-        result.chunks_deleted += context.store.delete_by_filter(&Filter::And(vec![
-            Filter::Eq("source_id".to_string(), context.config.source_id.clone()),
-            Filter::Eq("path".to_string(), path.to_string()),
-            Filter::Neq("seen_token".to_string(), context.sync_token.to_string()),
-        ]))?;
-        result.files_processed += 1;
-        result.files_processed_paths.push(path.to_string());
-
-        Ok(())
-    }
-}
-
-fn change_path(change: &FileChange) -> String {
-    match change {
-        FileChange::Added(path) | FileChange::Modified(path) | FileChange::Deleted(path) => {
-            path.clone()
-        }
-    }
-}
-
-fn empty_sync_result(dry_run: bool, already_up_to_date: bool) -> SyncResult {
-    SyncResult {
-        files_processed: 0,
-        chunks_added: 0,
-        chunks_updated: 0,
-        chunks_deleted: 0,
-        dry_run,
-        already_up_to_date,
-        files_processed_paths: Vec::new(),
-        elapsed_seconds: 0.0,
-        embedding_api_calls: 0,
-        embedded_texts: 0,
-        estimated_tokens: 0,
     }
 }
